+++ packages/typeorm/test/__fixture__/app/users/users.module.ts
@@ -7,6 +7,8 @@ import { UsersService } from './users.service';
 import { UsersCachedController } from './users-cached.controller';
 import { UsersCursorController } from './users-cursor.controller';
 import { UsersCursorNoLimitController } from './users-cursor-no-limit.controller';
+import { UsersCursorDefaultSortController } from './users-cursor-default-sort.controller';
+import { UsersCursorMultiSortController } from './users-cursor-multi-sort.controller';
 import { UsersController } from './users.controller';
 import { UsersWithStrategyController } from './users-with-strategy.controller';
 import { MeController } from './me.controller';
@@ -22,6 +24,8 @@ import { MeController } from './me.controller';
     UsersCachedController,
     UsersCursorController,
     UsersCursorNoLimitController,
+    UsersCursorDefaultSortController,
+    UsersCursorMultiSortController,
   ],
 })
 export class UsersModule {}